shifted general methods and properties too UI cls
